updating gemm kernel on TPUs and GPUs

+++ src/easydel/modules/phi/modelling_phi_flax.py
@@ -110,7 +110,7 @@ class FlaxPhiMLP(nn.Module):
 					blocksize_k=self.config.pallas_k_block_size,
 					blocksize_m=self.config.pallas_m_block_size,
 					blocksize_n=self.config.pallas_n_block_size,
-					po_dtype=self.dtype,
+					prod_dtype=self.dtype,
 					precision=self.precision,
 				),
 				in_axes=(0, None, None),
